steam/requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.6
django-ckeditor==6.1.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1